+++ R/group-indices.R
@@ -10,28 +10,38 @@
 #' @examples
 #' group_indices(mtcars, cyl)
 group_indices <- function(.data, ...) {
-  group_indices_(.data, .dots = lazyeval::lazy_dots(...))
+  UseMethod("group_indices")
 }
 
 #' @export
-#' @rdname group_indices
-group_indices_ <- function(.data, ..., .dots) {
+#' @rdname se-deprecated
+group_indices_ <- function(.data, ..., .dots = list()) {
   UseMethod("group_indices_")
 }
 
 #' @export
-group_indices_.data.frame <- function(.data, ..., .dots) {
-  .dots <- lazyeval::all_dots(..., .dots)
-  if (length(.dots) == 0L) {
+group_indices.data.frame <- function(.data, ...) {
+  dots <- tidy_quotes(...)
+  if (length(dots) == 0L) {
     return(rep(1L, nrow(.data)))
   }
-  grouped_indices_grouped_df_impl(group_by_(.data, .dots = .dots))
+  grouped_indices_grouped_df_impl(group_by(.data, !!! dots))
+}
+#' @export
+group_indices_.data.frame <- function(.data, ..., .dots = list()) {
+  dots <- compat_lazy_dots(.dots, caller_env(), .dots)
+  group_indices(.data, !!! dots)
 }
 
 #' @export
-group_indices_.grouped_df <- function(.data, ..., .dots) {
-  if (length(list(...)) || (!missing(.dots) && length(.dots))) {
-    warning("group_indices_.grouped_df ignores extra arguments")
+group_indices.grouped_df <- function(.data, ...) {
+  if (length(list(...))) {
+    warn("group_indices_.grouped_df ignores extra arguments")
   }
   grouped_indices_grouped_df_impl(.data)
+}
+#' @export
+group_indices_.grouped_df <- function(.data, ..., .dots = list()) {
+  dots <- compat_lazy_dots(.dots, caller_env(), .dots)
+  group_indices(.data, !!! dots)
 }