+++ R/lazy-ops.R
@@ -133,7 +133,7 @@ op_grps.op_summarise <- function(op) {
 
 #' @export
 op_grps.op_rename <- function(op) {
-  names(rename_vars_(op_grps(op$x), op$dots))
+  names(rename_vars(op_grps(op$x), !!! op$dots))
 }
 
 #' @export
@@ -162,11 +162,11 @@ op_vars.op_base <- function(op) {
 }
 #' @export
 op_vars.op_select <- function(op) {
-  names(select_vars_(op_vars(op$x), op$dots, include = op_grps(op$x)))
+  names(select_vars(op_vars(op$x), !!! op$dots, include = op_grps(op$x)))
 }
 #' @export
 op_vars.op_rename <- function(op) {
-  names(rename_vars_(op_vars(op$x), op$dots))
+  names(rename_vars(op_vars(op$x), !!! op$dots))
 }
 #' @export
 op_vars.op_summarise <- function(op) {