+++ R/rowwise.r
@@ -54,33 +54,41 @@ n_groups.rowwise_df <- function(x) {
 }
 
 #' @export
-group_by_.rowwise_df <- function(.data, ..., .dots, add = FALSE) {
-  warning("Grouping rowwise data frame strips rowwise nature", call. = FALSE)
+group_by.rowwise_df <- function(.data, ..., add = FALSE) {
+  warn("Grouping rowwise data frame strips rowwise nature")
   .data <- ungroup(.data)
 
-  groups <- group_by_prepare(.data, ..., .dots = .dots, add = add)
+  groups <- group_by_prepare(.data, ..., add = add)
   grouped_df(groups$data, groups$group_names)
+}
+#' @export
+group_by_.rowwise_df <- function(.data, ..., .dots = list(), add = FALSE) {
+  dots <- compat_lazy_dots(.dots, caller_env(), ...)
+  group_by(.data, !!! dots, add = add)
 }
 
 
 # Do ---------------------------------------------------------------------------
 
 #' @export
-do_.rowwise_df <- function(.data, ..., .dots) {
+do.rowwise_df <- function(.data, ...) {
   # Create ungroup version of data frame suitable for subsetting
   group_data <- ungroup(.data)
-
-  args <- lazyeval::all_dots(.dots, ...)
-  named <- named_args(args)
-  env <- new.env(parent = lazyeval::common_env(args))
   index <- attr(.data, "indices")
+
+  args <- tidy_quotes(...)
+  named <- named_args(args)
 
   # Create new environment, inheriting from parent, with an active binding
   # for . that resolves to the current subset. `_i` is found in environment
   # of this function because of usual scoping rules.
-  makeActiveBinding(".", function() {
-    lapply(group_data[`_i`, , drop = FALSE], "[[", 1)
-  }, env)
+  env <- child_env(NULL)
+  current_row <- function() lapply(group_data[`_i`, , drop = FALSE], "[[", 1)
+  env_assign_active(env, ".", current_row)
+  env_assign_active(env, ".data", current_row)
+
+  overscope <- new_overscope(env)
+  on.exit(overscope_clean(overscope))
 
   n <- nrow(.data)
   m <- length(args)
@@ -91,7 +99,7 @@ do_.rowwise_df <- function(.data, ..., .dots) {
 
   for (`_i` in seq_len(n)) {
     for (j in seq_len(m)) {
-      out[[j]][`_i`] <- list(eval(args[[j]]$expr, envir = env))
+      out[[j]][`_i`] <- list(overscope_eval(overscope, args[[j]]))
       p$tick()$print()
     }
   }
@@ -101,4 +109,9 @@ do_.rowwise_df <- function(.data, ..., .dots) {
   } else {
     label_output_list(NULL, out, groups(.data))
   }
+}
+#' @export
+do_.rowwise_df <- function(.data, ..., .dots = list()) {
+  dots <- compat_lazy_dots(.dots, caller_env(), ...)
+  do(.data, !!! dots)
 }