+++ R/sql-build.R
@@ -48,13 +48,13 @@ sql_build.op_base_local <- function(op, con, ...) {
 
 #' @export
 sql_build.op_select <- function(op, con, ...) {
-  vars <- select_vars_(op_vars(op$x), op$dots, include = op_grps(op$x))
+  vars <- select_vars(op_vars(op$x), !!! op$dots, include = op_grps(op$x))
   select_query(sql_build(op$x, con), ident(vars))
 }
 
 #' @export
 sql_build.op_rename <- function(op, con, ...) {
-  vars <- rename_vars_(op_vars(op$x), op$dots)
+  vars <- rename_vars(op_vars(op$x), !!! op$dots)
   select_query(sql_build(op$x, con), ident(vars))
 }