+++ R/ts.R
@@ -1,7 +1,4 @@
 #' @export
-filter_.ts <- function(.data, ..., .dots) {
-  stop(
-    "dplyr::filter() called with ts object. Do you want stats::filter()?",
-    call. = FALSE
-  )
+filter.ts <- function(.data, ...) {
+  abort("dplyr::filter() called with ts object. Do you want stats::filter()?")
 }