+++ tests/testthat/helper-astyle.R
@@ -20,12 +20,12 @@ astyle <- function(extra_args = character()) {
     "--align-reference=type"
   )
 
-  src_path <- normalizePath(vcapply(c("../../src", "../../inst/include"), testthat::test_path))
+  src_path <- normalizePath(map_chr(c("../../src", "../../inst/include"), testthat::test_path))
   src_files <- dir(src_path, "[.](?:cpp|h)$", recursive = TRUE, full.names = TRUE)
   astyle_files <- grep("(?:RcppExports[.]cpp|static_assert[.]h)", src_files, value = TRUE, invert = TRUE)
   output <- system2(astyle_cmd, c(astyle_args, astyle_files, extra_args), stdout = TRUE, stderr = TRUE)
   unchanged <- grepl("^Unchanged", output)
   if (any(!unchanged)) {
-    warning(paste(output[!unchanged], collapse = "\n"))
+    rlang::warn(paste(output[!unchanged], collapse = "\n"))
   }
 }