+++ tests/testthat/test-colwise.R
@@ -37,7 +37,7 @@ test_that("named arguments force complete namd", {
 })
 
 expect_classes <- function(tbl, expected) {
-  classes <- unname(vapply(tbl, class, character(1)))
+  classes <- unname(map_chr(tbl, class))
   classes <- paste0(substring(classes, 0, 1), collapse = "")
   expect_equal(classes, expected)
 }
@@ -69,24 +69,15 @@ test_that("empty selection does not select everything (#2009, #1989)", {
   expect_equal(mtcars, mutate_if(mtcars, is.factor, as.character))
 })
 
-test_that("arguments are matched by names and by position", {
-  fun <- function(a = 1, b = 2, c = 3) {
-    paste0(a, b, c)
-  }
-  call <- quote(fun())
-  expect_equal(eval(merge_args(call, list(a = 1))), "123")
-  expect_equal(eval(merge_args(call, list(c = 3, 1))), "123")
-  expect_equal(eval(merge_args(call, list(c = 3, 1, b = 2))), "123")
-
-  expect_equal(
-    as.list(mutate_all(mtcars, round, 0)),
-    lapply(mtcars, round, 0)
-  )
-})
-
 test_that("error is thrown with improper additional arguments", {
   expect_error(mutate_all(mtcars, round, 0, 0), "3 arguments passed")
-  expect_error(mutate_all(mtcars, mean, na.rm = TRUE, na.rm = TRUE), "multiple actual arguments")
+  expect_error(mutate_all(mtcars, mean, na.rm = TRUE, na.rm = TRUE), "Duplicate arguments")
+})
+
+test_that("fun_list is merged with new args", {
+  funs <- funs(fn = bar)
+  funs <- as_fun_list(funs, baz = "baz")
+  expect_identical(funs$fn, ~bar(., baz = "baz"))
 })
 
 
@@ -95,11 +86,11 @@ test_that("error is thrown with improper additional arguments", {
 test_that("_each() and _all() families agree", {
   df <- data.frame(x = 1:3, y = 1:3)
 
-  expect_equal(summarise_each(df, funs(mean)), summarise_all(df, mean))
-  expect_equal(summarise_each(df, funs(mean), x:y), summarise_at(df, vars(x:y), mean))
-  expect_equal(summarise_each(df, funs(mean), z = y), summarise_at(df, vars(z = y), mean))
+  expect_warning(expect_equal(summarise_each(df, funs(mean)), summarise_all(df, mean)), "deprecated")
+  expect_warning(expect_equal(summarise_each(df, funs(mean), x:y), summarise_at(df, vars(x:y), mean)), "deprecated")
+  expect_warning(expect_equal(summarise_each(df, funs(mean), z = y), summarise_at(df, vars(z = y), mean)), "deprecated")
 
-  expect_equal(mutate_each(df, funs(mean)), mutate_all(df, mean))
-  expect_equal(mutate_each(df, funs(mean), x:y), mutate_at(df, vars(x:y), mean))
-  expect_equal(mutate_each(df, funs(mean), z = y), mutate_at(df, vars(z = y), mean))
+  expect_warning(expect_equal(mutate_each(df, funs(mean)), mutate_all(df, mean)), "deprecated")
+  expect_warning(expect_equal(mutate_each(df, funs(mean), x:y), mutate_at(df, vars(x:y), mean)), "deprecated")
+  expect_warning(expect_equal(mutate_each(df, funs(mean), z = y), mutate_at(df, vars(z = y), mean)), "deprecated")
 })