+++ tests/testthat/test-lazy-ops.R
@@ -114,26 +114,24 @@ test_that("unsorted gives NULL", {
 })
 
 test_that("arranges captures DESC", {
-  out <- lazy_frame(x = 1:3, y = 3:1) %>%
-    arrange(desc(x))
+  out <- lazy_frame(x = 1:3, y = 3:1) %>% arrange(desc(x))
 
-  expect_equal(op_sort(out), list(quote(desc(x))))
+  expect_equal(op_sort(out), list(~desc(x)))
 })
 
 test_that("multiple arranges combine", {
-  out <- lazy_frame(x = 1:3, y = 3:1) %>%
-    arrange(x) %>%
-    arrange(y)
+  out <- lazy_frame(x = 1:3, y = 3:1) %>% arrange(x) %>% arrange(y)
+  out <- arrange(arrange(lazy_frame(x = 1:3, y = 3:1), x), y)
 
-  expect_equal(op_sort(out), list(quote(x), quote(y)))
+  expect_equal(op_sort(out), list(~x, ~y))
 })
 
 test_that("preserved across compute and collapse", {
   df1 <- memdb_frame(x = sample(10)) %>% arrange(x)
 
   df2 <- compute(df1)
-  expect_equal(op_sort(df2), list(quote(x)))
+  expect_equal(op_sort(df2), list(~x))
 
   df3 <- collapse(df1)
-  expect_equal(op_sort(df3), list(quote(x)))
+  expect_equal(op_sort(df3), list(~x))
 })